Added default blueprint
Blueprint uses `.addBowerPackageToProject` and `.addAddonToProject` to include `ember-simple-auth` and `ember-cli-simple-auth` respectively.

In theory, instead of using `npm install` or `bower install`, we're supposed to use `ember install:addon ember-gdrive` and the default blueprint will run with the installation, pulling everything that's required. However, I didn't get the chance to test this yet.

It may be that we require a newer version of ember-CLI, or that i the addon first needs to be published to ember-addons. I will investigate this next.