Added document-creator component
* required adding a new method to DocumentSource
* required adding a new action and method to ApplicationRouteMixin - also did some refactoring
* required injecting documentSource into new component - also did some initializer refactoring

On the side of the app, pretty much any template can use the component and bind it to an action, but only the `ApplicationRouteMixin` already has a predefined action called `documentCreated`.To use it there, we add `{{document-creator documentCreated="documentCreated"}}`. The action has one parameter, which is the created document, so the typical response would be to transition to the `document` route with the new document as the model - `route.transitionTo('document', doc)`.