Applied original storypad styles to share-modal
I've copied over all of the styling rules that used to be applied to the share-modal while it was still in storypad.
They are all organized under the .share-modal class and defined in such a way to make the selectors less specific is possible, so they can easily be overriden in the app that uses the component. I wasn't too thorough, though, since I believe that should be a separate task.