Removed ember-cli-simple-auth dependency.
Due to the fact that `ember-cli-simple-auth` doesn't make use of the new `parentAddon` feature in ember-cli add-ons, there is no use having it as an ember-gdrive dependency. It actually just needs to be installed to the parent app together with ember-gdrive via the default blueprint instead.

If there's ever an update to ember-cli-simple-auth that would make it work via nesting, then we can move it back here, but for now, it's completely redundant and does nothing.