Added scaffolding for basic app functionality to default blueprint
The way this works is that everything within the blueprint's `files` folder gets merged with the parent projects file structure. Luckily, unlike the `insertIntoFile` functionality, this is far more robust and actually asks for user to confirm overwriting files already present.

With these changes, when `ember-gdrive` is installed via `ember install:addon`, a set of routes, templates and controllers will be created that supports creating a new google document, auto-navigating to it and then adding to and removing from a list of "items". The interface is simple and there's no styling, but the functionality is present.